Move ints into a dir
