+++ crates/cairo-lang-sierra-to-casm/src/invocations/int/unsigned256.rs
@@ -1,11 +1,11 @@
 use cairo_lang_casm::builder::CasmBuilder;
 use cairo_lang_casm::casm_build_extend;
-use cairo_lang_sierra::extensions::uint256::Uint256Concrete;
+use cairo_lang_sierra::extensions::int::unsigned256::Uint256Concrete;
 use num_bigint::BigInt;
 
-use super::{CompiledInvocation, CompiledInvocationBuilder, InvocationError};
 use crate::invocations::{
-    add_input_variables, get_non_fallthrough_statement_id, CostValidationInfo,
+    add_input_variables, get_non_fallthrough_statement_id, CompiledInvocation,
+    CompiledInvocationBuilder, CostValidationInfo, InvocationError,
 };
 
 /// Builds instructions for Sierra u256 operations.
