+++ crates/cairo-lang-sierra/src/extensions/core.rs
@@ -11,13 +11,18 @@ use super::ec::{EcLibfunc, EcOpType, EcPointType, EcStateType};
 use super::enm::{EnumLibfunc, EnumType};
 use super::felt252_dict::{Felt252DictLibfunc, Felt252DictType};
 use super::gas::BuiltinCostsType;
+use super::int::unsigned::{
+    Uint16Libfunc, Uint16Type, Uint32Libfunc, Uint32Type, Uint64Libfunc, Uint64Type, Uint8Libfunc,
+    Uint8Type,
+};
+use super::int::unsigned128::{Uint128Libfunc, Uint128Type};
+use super::int::unsigned256::Uint256Libfunc;
 use super::modules::boxing::{BoxLibfunc, BoxType};
 use super::modules::felt252::{Felt252Libfunc, Felt252Type};
 use super::modules::function_call::FunctionCallLibfunc;
 use super::modules::gas::{GasBuiltinType, GasLibfunc};
 use super::modules::mem::MemLibfunc;
 use super::modules::non_zero::{NonZeroType, UnwrapNonZeroLibfunc};
-use super::modules::uint128::{Uint128Libfunc, Uint128Type};
 use super::modules::unconditional_jump::UnconditionalJumpLibfunc;
 use super::nullable::{NullableLibfunc, NullableType};
 use super::pedersen::{PedersenLibfunc, PedersenType};
@@ -28,11 +33,6 @@ use super::snapshot::{SnapshotTakeLibfunc, SnapshotType};
 use super::squashed_felt252_dict::SquashedFelt252DictType;
 use super::starknet::{StarkNetLibfunc, StarkNetType};
 use super::structure::{StructLibfunc, StructType};
-use super::uint::{
-    Uint16Libfunc, Uint16Type, Uint32Libfunc, Uint32Type, Uint64Libfunc, Uint64Type, Uint8Libfunc,
-    Uint8Type,
-};
-use super::uint256::Uint256Libfunc;
 use super::uninitialized::UninitializedType;
 use crate::{define_libfunc_hierarchy, define_type_hierarchy};
