+++ crates/cairo-lang-sierra/src/extensions/modules/casts.rs
@@ -1,15 +1,14 @@
 use cairo_lang_utils::unordered_hash_map::UnorderedHashMap;
 
+use super::int::unsigned::{Uint16Type, Uint32Type, Uint64Type, Uint8Type};
+use super::int::unsigned128::Uint128Type;
 use super::range_check::RangeCheckType;
-use super::uint::Uint16Type;
 use crate::define_libfunc_hierarchy;
 use crate::extensions::lib_func::{
     BranchSignature, DeferredOutputKind, LibfuncSignature, OutputVarInfo, ParamSignature,
     SierraApChange, SignatureOnlyGenericLibfunc, SignatureSpecializationContext,
     SpecializationContext,
 };
-use crate::extensions::uint::{Uint32Type, Uint64Type, Uint8Type};
-use crate::extensions::uint128::Uint128Type;
 use crate::extensions::{
     args_as_two_types, NamedLibfunc, NamedType, OutputVarReferenceInfo,
     SignatureBasedConcreteLibfunc, SpecializationError,
