+++ crates/cairo-lang-sierra/src/extensions/modules/int/mod.rs
@@ -1,0 +1,10 @@
+pub mod unsigned;
+pub mod unsigned128;
+pub mod unsigned256;
+
+/// Operators for integers.
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum IntOperator {
+    OverflowingAdd,
+    OverflowingSub,
+}
