+++ crates/cairo-lang-sierra/src/extensions/modules/starknet/getter.rs
@@ -5,10 +5,10 @@ use super::interoperability::ContractAddressType;
 use super::syscalls::SyscallGenericLibfunc;
 use crate::extensions::boxing::BoxType;
 use crate::extensions::felt252::Felt252Type;
+use crate::extensions::int::unsigned::Uint64Type;
+use crate::extensions::int::unsigned128::Uint128Type;
 use crate::extensions::lib_func::SignatureSpecializationContext;
 use crate::extensions::structure::StructType;
-use crate::extensions::uint::Uint64Type;
-use crate::extensions::uint128::Uint128Type;
 use crate::extensions::{NamedType, NoGenericArgsGenericType, SpecializationError};
 use crate::ids::{ConcreteTypeId, UserTypeId};
 use crate::program::GenericArg;
